SignedQuery should respect explicity body and content-type choices
A PostQuery used to override any explicityly given Body and
Content-Type choices in the SignedQuery. I belive this is to be
erroneous, as it simply overrode explicit selections due to a
PostQuery, which is particularly impossible to deal with when
implementing AWS version 4 signing.

It now enforces the PostQuery defaults *only* if the explicit
selections are missing.