Minor tweak to Aws.Core.queryToHttpRequest to support Elastic Transcoder
The Elastic Transcoder API insists that simplified 8601 dates be used
for its Date: header. To get arround this the Elastic Transcoder suite
puts an apropriately formatted date header into the headers that gets
passed to queryToHttprequest. This patch modifies queryToHttprequest
to see if any Date: binding is present in the headers passed to it;
if such a binding is present then merely refrains from adding a
Date: header and uses the one passed to it.