update to conduit-1.3 and http-conduit-2.3
The new conduit removed ResumableSource, replacing it with
SealedConduitT. Rather than use that where it is not needed,
changed to using runConduit and .| where it used to use $$+-

The removal of ResumableSource required the removal of this line from
s3ResponseConsumer:
            C.closeResumableSource (HTTP.responseBody resp)
So there's a potential for a change to the http response resource
lifetime having been introduced by these changes. That line was added
in 6b9c070 as part of a memory leak in S3 multipart upload fix for
issue #142. I have verified that Examples/MultipartUpload.hs uses
the same amount of memory after this change as it did before.

Several parts of this package's API used ResumableSource, and with
its removal now use ResourceT. Also, http-conduit-2.3.0 removed withManager.
These changes will affect users of the API, where they used to have eg:

	withManager $ \mgr -> Aws.pureAws cfg s3cfg mgr ...

Now they will need:

	mgr <- newManager tlsManagerSettings
	runResourceT $ Aws.pureAws cfg s3cfg mgr ...

Code and examples have been updated accordingly; this will need
to be mentioned in the changelog and would be a good reason for a PVP
major version bump.

Also, added a dependency on exceptions, since ResourceT
no longer has a MonadBaseControl instance.

Thanks to Dan Burton for his help with this.