support network-3
fixes https://github.com/aristidb/aws/issues/264

This keeps support for building with network-2 by adding a build flag.
Cabal will automatically disable the build flag if that's the only way
to satisfy the dependencies.

At some point in the future when network-2 compatability is not useful,
it would make sense to remove the build flag.