convert deprecated conduit operators
Avoids some build warnings.

$= and =$= are the same and .| replaces them

The $$ deprecation message is "Use runConduit and .|" but I'm
not clear how to as a non-user of conduit. But I noticed that
conduit still contains connect = ($$) which is not deprecated.
So use that. Unfortunately, it has a different precidence than
the $$ operator, so have to add parens.