support PutObject tagging
Note that the x-amz-tagging header is only sent when poTagging is not
empty, because a specific permission is needed to use the header, which
not all buckets may have.

This will need a new major version since it changes constructors.

closes https://github.com/aristidb/aws/issues/294
(although more could be added later to support the tagging API)